Start adding websocket support: - send test message between client and servers - move http logic into its own module - properly export GameDO in new http module - add WS proxy for local development. Add env vars   for WEBSOCKET_URL
